Use NSInteger/NSUInteger
They are ugly, but we can’t fight the future.

+++ NSNumber+Ruby.m
@@ -3,10 +3,10 @@
 
 @implementation NSNumber (Ruby)
 
-- (NSArray *(^)(int))upTo {
-    return ^NSArray *(int upTo) {
-        int const start = self.intValue;
-        int const count = upTo - start + 1;
+- (NSArray *(^)(NSInteger))upTo {
+    return ^NSArray *(NSInteger upTo) {
+        NSInteger const start = self.intValue;
+        NSInteger const count = upTo - start + 1;
         id objs[count];
         for (int x = 0; x < count; ++x)
             objs[x] = @(start + x);
